# openapi.yaml
openapi: "3.0.1"
info:
  title: API KV
  version: "1"
  description: |
    Key-Value simple storage with REST API, focused on storage capacity, not on latency.
servers:
  - url: /
security:
  - HeaderAuth: [ ]
paths:
  /{namespace}:
    get:
      operationId: keys
      summary: List keys in namespace.
      description: |
        List keys in namespace page by page. Iteration may not be consistent if there are updates between pages.
        
        At the end of page, operation will return empty list.
        Iteration order is non-deterministic, but generally tends to be from oldest key to newest key.
      parameters:
        - in: path
          name: namespace
          description: Key-Value namespace
          required: true
          schema:
            type: string
        - in: query
          name: cursor
          description: Cursor of the previous page, should be passed in from the previous request or absent for new request.
          required: false
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/batch"

    delete:
      operationId: deleteNamespace
      summary: Delete namespace and all keys
      parameters:
        - in: path
          name: namespace
          description: Key-Value namespace
          required: true
          schema:
            type: string
      responses:
        204:
          description: OK
...
components:
  schemas:
    batch:
      type: object
      properties:
        keys:
          description: Keys list
          type: array
          items:
            type: string

        cursor:
          description: Cursor used for pagination
          type: integer
          format: int64
      required:
        - keys
        - cursor
  securitySchemes:
    HeaderAuth:
      type: apiKey
      in: header
      name: X-Api-Key